Conflicts:
	src/maidsafe/vault/data_manager/dispatcher.h
	src/maidsafe/vault/operation_handlers.cc
	src/maidsafe/vault/operation_handlers.h

+++ src/maidsafe/vault/maid_manager/metadata.h
@@ -80,10 +80,11 @@ MaidManagerMetadata::Status MaidManagerMetadata::AllowPut(const Data& data) cons
   for (const auto& pmid_total : pmid_totals_)
     total_claimed_available_size_by_pmids += pmid_total.pmid_metadata.claimed_available_size;
   auto cost(data.Serialise()->string().size());
-  if (total_claimed_available_size_by_pmids < (total_put_data_ + cost))
+  if (total_claimed_available_size_by_pmids < (static_cast<int64_t>(total_put_data_ + cost)))
     return Status::kNoSpace;
 
-  return ((3 * total_claimed_available_size_by_pmids / 100) < (total_put_data_ + cost))
+  return ((3 * total_claimed_available_size_by_pmids / 100) <
+               static_cast<int64_t>(total_put_data_ + cost))
              ? Status::kLowSpace
              : Status::kOk;
 }
